Changed folder structure; Added Image
For Menu and subMenus, I've changed the folder distribution.

I've also added an Image for taxis and appropriate folder for it.

Loop through images may be messy since it may require an animation.

For different menu screens, we may be able to overlay them so long as we
do ISwitchable stuff to it.